Move HPPC related classes to o.a.l.hppc package Also forked indexRemove method
